create resource table component
